pyproject.toml
[project]
name = "discord-utils"
version = "0.1.0"
description = "Shared Discord utilities for the Iron Foundry clan."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.7.1",
    "loguru>=0.7.3",
    "pillow>=12.1.1",
    "pydantic>=2.12.3",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.2",
]
